added rollup to generate dist, changed resolver ABI to v3
