+++ package.json
@@ -1,12 +1,17 @@
 {
   "name": "@linagee/lnr-ethers",
-  "version": "1.2.1",
+  "version": "1.3.0-rc.0",
   "description": "Integrates the LNR resolver and Ethers in an easy to use package",
-  "main": "./dist/index.js",
+  "main": "dist/cjs/index.cjs.js",
+  "module": "dist/es/index.es.js",
+  "browser": "dist/umd/index.umd.js",
+  "unpkg": "dist/umd-min/index.umd.min.js",
   "scripts": {
     "production": "WEBPACK_ENV=build webpack",
     "development": "WEBPACK_ENV=dev webpack --progress --colors --watch",
-    "changelog": "auto-changelog -p && git add CHANGELOG.md"
+    "changelog": "auto-changelog -p && git add CHANGELOG.md",
+    "clean": "rm -rf dist",
+    "build": "npm run clean && rollup -c"
   },
   "files": [
     "./dist/*"
@@ -17,6 +22,10 @@
     "ethers"
   ],
   "author": "Derp Herpenstein",
+  "contributors": [
+    "Chriton",
+    "Mason Keresty"
+  ],
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -31,7 +40,14 @@
     "ethers": "^5.7.2"
   },
   "devDependencies": {
+    "@babel/core": "^7.21.4",
+    "@babel/preset-env": "^7.21.4",
+    "@rollup/plugin-terser": "^0.4.1",
     "auto-changelog": "^2.4.0",
+    "rollup": "^2.79.1",
+    "rollup-plugin-babel": "^4.4.0",
+    "rollup-plugin-commonjs": "^10.1.0",
+    "rollup-plugin-node-resolve": "^5.2.0",
     "webpack-cli": "^5.0.1"
   }
 }
